+++ cluster.go
@@ -118,6 +118,7 @@ type Cluster struct {
 	Nodes         []*Node  `json:"nodes"          yaml:"nodes"`
 	SSHKey        string   `json:"ssh_key"        yaml:"ssh_key"`
 	ServiceSubnet string   `json:"service_subnet" yaml:"service_subnet"`
+	PodSubnet     string   `json:"pod_subnet"     yaml:"pod_subnet"`
 	DNSServers    []string `json:"dns_servers"    yaml:"dns_servers"`
 	Options       Options  `json:"options"        yaml:"options"`
 }
@@ -129,6 +130,10 @@ func (c *Cluster) Validate() error {
 	}
 
 	_, _, err := net.ParseCIDR(c.ServiceSubnet)
+	if err != nil {
+		return err
+	}
+	_, _, err = net.ParseCIDR(c.PodSubnet)
 	if err != nil {
 		return err
 	}