+++ cluster_test.go
@@ -22,6 +22,7 @@ nodes:
       label1: value1
 ssh_key: clusterkey
 service_subnet: 12.34.56.00/24
+pod_subnet: 10.1.0.0/16
 dns_servers: ["1.1.1.1", "8.8.8.8"]
 options:
   etcd:
@@ -90,6 +91,9 @@ options:
 	if c.ServiceSubnet != "12.34.56.00/24" {
 		t.Error(`c.ServiceSubnet != "12.34.56.00/24"`)
 	}
+	if c.PodSubnet != "10.1.0.0/16" {
+		t.Error(`c.PodSubnet != "10.1.0.0/16"`)
+	}
 	if !reflect.DeepEqual(c.DNSServers, []string{"1.1.1.1", "8.8.8.8"}) {
 		t.Error(`!reflect.DeepEqual(c.DNSServers, []string{"1.1.1.1", "8.8.8.8"})`)
 	}
@@ -130,6 +134,7 @@ func testClusterValidate(t *testing.T) {
 			Cluster{
 				Name:          "",
 				ServiceSubnet: "10.0.0.0/14",
+				PodSubnet:     "10.1.0.0/16",
 			},
 			true,
 		},
@@ -138,6 +143,16 @@ func testClusterValidate(t *testing.T) {
 			Cluster{
 				Name:          "testcluster",
 				ServiceSubnet: "",
+				PodSubnet:     "10.1.0.0/16",
+			},
+			true,
+		},
+		{
+			"No pod subnet",
+			Cluster{
+				Name:          "testcluster",
+				ServiceSubnet: "10.1.0.0/16",
+				PodSubnet:     "",
 			},
 			true,
 		},
@@ -146,6 +161,7 @@ func testClusterValidate(t *testing.T) {
 			Cluster{
 				Name:          "testcluster",
 				ServiceSubnet: "10.0.0.0/14",
+				PodSubnet:     "10.1.0.0/16",
 				DNSServers:    []string{"a.b.c.d"},
 			},
 			true,
@@ -155,6 +171,7 @@ func testClusterValidate(t *testing.T) {
 			Cluster{
 				Name:          "testcluster",
 				ServiceSubnet: "10.0.0.0/14",
+				PodSubnet:     "10.1.0.0/16",
 			},
 			false,
 		},