+++ kubernetes_strategy_test.go
@@ -1,6 +1,7 @@
 package cke
 
 import (
+	"path/filepath"
 	"reflect"
 	"strings"
 	"testing"
@@ -169,6 +170,12 @@ func testKubernetesDecideToDo(t *testing.T) {
 				Rivers: allNodes, APIServers: cpNodes, ControllerManagers: cpNodes, Schedulers: cpNodes,
 			},
 			Commands: []Command{
+				{"image-pull", ToolsImage.Name(), ""},
+				{"mkdir", cniBinDir, ""},
+				{"mkdir", cniConfDir, ""},
+				{"mkdir", cniVarDir, ""},
+				{"make-file", filepath.Join(cniConfDir, "98-bridge.conf"), ""},
+				{"run-container", strings.Join(allNodes, ","), "install-cni"},
 				{"image-pull", HyperkubeImage.Name(), ""},
 				{"mkdir", "/var/log/kubernetes/kubelet", ""},
 				{"mkdir", "/var/log/pods", ""},