+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Opt in to [Go modules](https://github.com/golang/go/wiki/Modules).
+- Support for RBAC.
+- Enable CNI plugin.
 
 ## [0.5] - 2018-09-13